+++ common/client-libs/validator-client/src/nyxd/contract_traits/mod.rs
@@ -10,6 +10,7 @@ use std::str::FromStr;
 // query clients
 mod coconut_bandwidth_query_client;
 mod dkg_query_client;
+mod ephemera_query_client;
 mod group_query_client;
 mod mixnet_query_client;
 mod multisig_query_client;
@@ -20,6 +21,7 @@ mod vesting_query_client;
 // signing clients
 mod coconut_bandwidth_signing_client;
 mod dkg_signing_client;
+mod ephemera_signing_client;
 mod group_signing_client;
 mod mixnet_signing_client;
 mod multisig_signing_client;
@@ -32,6 +34,7 @@ pub use coconut_bandwidth_query_client::{
     CoconutBandwidthQueryClient, PagedCoconutBandwidthQueryClient,
 };
 pub use dkg_query_client::{DkgQueryClient, PagedDkgQueryClient};
+pub use ephemera_query_client::{EphemeraQueryClient, PagedEphemeraQueryClient};
 pub use group_query_client::{GroupQueryClient, PagedGroupQueryClient};
 pub use mixnet_query_client::{MixnetQueryClient, PagedMixnetQueryClient};
 pub use multisig_query_client::{MultisigQueryClient, PagedMultisigQueryClient};
@@ -42,6 +45,7 @@ pub use vesting_query_client::{PagedVestingQueryClient, VestingQueryClient};
 // re-export signing traits
 pub use coconut_bandwidth_signing_client::CoconutBandwidthSigningClient;
 pub use dkg_signing_client::DkgSigningClient;
+pub use ephemera_signing_client::EphemeraSigningClient;
 pub use group_signing_client::GroupSigningClient;
 pub use mixnet_signing_client::MixnetSigningClient;
 pub use multisig_signing_client::MultisigSigningClient;
@@ -61,6 +65,9 @@ pub trait NymContractsProvider {
     fn group_contract_address(&self) -> Option<&AccountId>;
     fn multisig_contract_address(&self) -> Option<&AccountId>;
 
+    // ephemera-related
+    fn ephemera_contract_address(&self) -> Option<&AccountId>;
+
     // SPs
     fn name_service_contract_address(&self) -> Option<&AccountId>;
     fn service_provider_contract_address(&self) -> Option<&AccountId>;
@@ -75,6 +82,8 @@ pub struct TypedNymContracts {
     pub group_contract_address: Option<AccountId>,
     pub multisig_contract_address: Option<AccountId>,
     pub coconut_dkg_contract_address: Option<AccountId>,
+
+    pub ephemera_contract_address: Option<AccountId>,
 
     pub service_provider_directory_contract_address: Option<AccountId>,
     pub name_service_contract_address: Option<AccountId>,
@@ -107,6 +116,10 @@ impl TryFrom<NymContracts> for TypedNymContracts {
                 .transpose()?,
             coconut_dkg_contract_address: value
                 .coconut_dkg_contract_address
+                .map(|addr| addr.parse())
+                .transpose()?,
+            ephemera_contract_address: value
+                .ephemera_contract_address
                 .map(|addr| addr.parse())
                 .transpose()?,
             service_provider_directory_contract_address: value