+++ common/client-libs/validator-client/src/nyxd/mod.rs
@@ -259,6 +259,10 @@ impl<C, S> NymContractsProvider for NyxdClient<C, S> {
         self.config.contracts.multisig_contract_address.as_ref()
     }
 
+    fn ephemera_contract_address(&self) -> Option<&AccountId> {
+        self.config.contracts.ephemera_contract_address.as_ref()
+    }
+
     fn name_service_contract_address(&self) -> Option<&AccountId> {
         self.config.contracts.name_service_contract_address.as_ref()
     }