+++ ephemera/src/core/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod api_cmd;
+pub(crate) mod builder;
+pub(crate) mod ephemera;
+pub(crate) mod shutdown;